14177: DAC optimization of DP
Always look for reuse windows

Apporach for cost function was right, but precomputing under the
assumptoin that some cost function values might be reused consumed
unnecessary time. First attempt resulted in TLE.

Observe that when interation in compute function is performed backwads,
to compute the next cost, you only have to add the partial sum of the
next variable that is going to be included in the set. This reduces the
number of additions performed from N^2 to N.