15586, 15591: Disjoint sets with offline query
When a video is connected, its similarity is the minimum edge weight
among all edges connecting A to B.
Sort edges with decreasing weight, sort queries with decreasing
similarity.
Build disjoint set with that order and check size of connected
components when all edges with weight bigger than K is added.